Bug fix. Ignore cached files. Custom board sizes.
Whenever ever cell on the board was flagged, the game would end saying
the user had won.
I added the pycache directory to the .gitignore so it wouldn't upload
that folder.
Board sizes are no longer fixed (although it was mostly true to the
original implementation of the game, I figured people would want to make
their own board sizes) and the number of mines is also customizable
(with a cap at half the volume of the board).